# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Роботоподбор API
  version: 0.1.0
  description: >
    Стаб контракта для фронтенда. Бизнес-логика расчёта живёт на бэкенде,
    фронт только отображает готовые значения.
  license:
    name: Proprietary
servers:
  - url: /api
security:
  - sessionCookie: []
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: Проекты пользователя
      description: Только для авторизованного пользователя. Без limit возвращается весь список.
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Расчёты пользователя
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
        '401':
          description: Требуется вход
    post:
      operationId: createProject
      summary: Создать проект
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '401':
          description: Требуется вход
...
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Проект
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Проект с входными параметрами и сценариями
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '404':
          description: Не найден
    patch:
      operationId: updateProject
      summary: Изменить проект
      description: Переименование, смена статуса, правка паспорта объекта (ТЗ 3.1.3).
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectPatch' }
      responses:
        '200':
          description: Проект обновлён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '400':
          description: Некорректные данные
        '404':
          description: Не найден
    delete:
      operationId: deleteProject
      summary: Удалить проект
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Удалён
        '401': { description: Требуется вход }
        '404': { description: Не найден }
  /projects/{projectId}/calculations:
    get:
      operationId: listProjectCalculations
      summary: История расчётов проекта
      description: История сохранённых расчётов проекта, новые сверху (ТЗ 3.1.5).
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Снимки расчётов с версиями данных и модели
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CalculationSnapshot' }
        '404':
          description: Не найден
    post:
      operationId: saveProjectCalculation
      summary: Сохранить снимок расчёта
      description: >
        Сохраняет снимок расчёта: результат целиком, паспорт объекта на момент расчёта,
        версию данных (каталог и нормативы) и версию расчётной модели. Повторное открытие
        снимка воспроизводит тот же результат без пересчёта.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalculationSnapshotInput' }
      responses:
        '201':
          description: Снимок сохранён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalculationSnapshot' }
        '400':
          description: Некорректные данные
        '404':
          description: Не найден
  /projects/{projectId}/copy:
    post:
      operationId: copyProject
      summary: Копировать проект
      description: >
        Создаёт копию проекта для сравнения сценариев внутри проекта (см. 3.1.3 ТЗ:
        не менее трёх сценариев). Доступно только владельцу проекта.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Копия создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '401':
          description: Требуется вход
        '404':
          description: Не найден
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: httpOnly-cookie сессии, выдаётся при входе и регистрации.
  schemas:
...
    ProjectInput:
      type: object
      required: [title, objectType, parameters]
      properties:
        title: { type: string }
        objectType: { type: string }
        parameters:
          type: object
          additionalProperties: { type: string }
        processes:
          type: array
          items: { type: string }
        solutionId:
          type: string
          description: Решение каталога, выбранное на шаге подбора.
        calculation:
          $ref: '#/components/schemas/CalculationResult'
        modelVersion:
          type: string
          description: Версия расчётной модели, выдавшей результат.
...
    ProjectPatch:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { $ref: '#/components/schemas/Maturity' }
        parameters:
          type: object
          additionalProperties: { type: string }
        processes:
          type: array
          items: { type: string }
        solutionId: { type: string }
    CalculationSnapshotInput:
      type: object
      required: [calculation]
      properties:
        calculation: { $ref: '#/components/schemas/CalculationResult' }
        solutionId: { type: string }
        modelVersion: { type: string }
    CalculationSnapshot:
      type: object
      required: [id, createdAt, dataVersion, modelVersion]
      properties:
        id:
          type: string
          description: Идентификатор снимка; его же принимает GET /calculations/{calculationId}.
        createdAt: { type: string, format: date-time }
        dataVersion:
          type: string
          description: Версия каталога и нормативов на момент расчёта (data-YYYYMMDDHHMMSS).
        modelVersion: { type: string }
        solutionId: { type: string }
        payback: { type: string }
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          required: [objectType, parameters, scenarios]
          properties:
            objectType: { type: string }
            parameters:
              type: object
              additionalProperties: { type: string }
            processes:
              type: array
              items: { type: string }
            scenarios:
              type: array
              description: Минимум три — «как есть», «покупка», «услуга/RaaS».
              items: { $ref: '#/components/schemas/ScenarioBar' }
            calculationId: { type: string }
            solutionId: { type: string }
            dataVersion: { type: string }
            modelVersion: { type: string }
            calculatedAt: { type: string, format: date-time }
...
    SensitivityFactor:
      type: object
      required: [id, label, impact, direction]
      properties:
        id: { type: string }
        label: { type: string }
        impact:
          type: number
          description: Влияние на срок окупаемости, доля от базового значения.
        direction:
          type: string
          enum: [up, down]
        low: { type: string }
        high: { type: string }
    DataConfidence:
      type: string
      enum: [confirmed, needs-review]
    Maturity:
      type: string
      enum: [operation, piloting, rnd]
...
    ScenarioBar:
      type: object
      required: [id, title, subtitle, tco, share, delta]
      properties:
        id: { type: string }
        title: { type: string }
        subtitle: { type: string }
        tco: { type: number }
        share: { type: number }
        delta: { type: string }
        detail: { type: string }
        recommended: { type: boolean }
    CostLine:
      type: object
      required: [title, amount, share, source]
      properties:
        title: { type: string }
        amount: { type: number }
        share: { type: integer }
        source: { type: string }
        confidence: { $ref: '#/components/schemas/DataConfidence' }
    CostGroup:
      type: object
      required: [id, title, amount, share, confidence, lines]
      properties:
        id: { type: string }
        title: { type: string }
        amount: { type: number }
        share: { type: integer }
        confidence: { $ref: '#/components/schemas/DataConfidence' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/CostLine' }
    CalculationResult:
      type: object
      required:
        [
          id, objectTitle, meta, payback, capex, roi, scenarios, costGroups,
          opexSaving, warning, assumptions, totalTco,
        ]
      properties:
        id: { type: string }
        objectTitle: { type: string }
        meta: { type: string }
        payback: { $ref: '#/components/schemas/Kpi' }
        capex: { $ref: '#/components/schemas/Kpi' }
        roi: { $ref: '#/components/schemas/Kpi' }
        opexSaving:
          type: object
          required: [series, percent, meta]
          properties:
            series: { type: array, items: { type: number } }
            percent: { type: string }
            meta: { type: string }
        warning: { type: string }
        assumptions: { type: array, items: { type: string } }
        scenarios:
          type: array
          items: { $ref: '#/components/schemas/ScenarioBar' }
        costGroups:
          type: array
          items: { $ref: '#/components/schemas/CostGroup' }
        totalTco: { type: number }
        dataVersion:
          type: string
          description: Версия каталога и нормативов, на которых сделан расчёт (ТЗ 3.1.5).
        modelVersion:
          type: string
          description: Версия расчётной модели.
        calculatedAt: { type: string, format: date-time }
        sensitivity:
          type: array
          description: Параметры, ранжированные по влиянию на результат.
          items: { $ref: '#/components/schemas/SensitivityFactor' }
    Kpi:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
        unit: { type: string }
        note: { type: string }
        trend: { type: string, enum: [up, down, none] }
...
    Project:
      type: object
      required: [id, title, meta, payback, status]
      properties:
        id: { type: string }
        title: { type: string }
        meta: { type: string }
        payback: { type: string }
        status: { $ref: '#/components/schemas/Maturity' }